a days worth of pulling my hair out
