+++ packages/assembly/assembly/internal/TestNodeType.ts
@@ -1,0 +1,6 @@
+export const enum TestNodeType {
+  /** A Test. */
+  Test = 0,
+  /** A group. */
+  Group = 1,
+}
