+++ packages/core/src/index.ts
@@ -1,18 +1,10 @@
-import CombinationReporter from "./reporter/CombinationReporter";
-import EmptyReporter from "./reporter/EmptyReporter";
-import SummaryReporter from "./reporter/SummaryReporter";
-import VerboseReporter from "./reporter/VerboseReporter";
-
-export { CombinationReporter };
-export { EmptyReporter };
-export { SummaryReporter };
-export { VerboseReporter };
+export * from "./reporter/CombinationReporter";
+export * from "./reporter/EmptyReporter";
+export * from "./reporter/IReporter";
+export * from "./reporter/SummaryReporter";
+export * from "./reporter/VerboseReporter";
 export * from "./test/IWarning";
 export * from "./test/TestContext";
-export * from "./test/TestGroup";
-export * from "./test/TestReporter";
-export * from "./test/TestResult";
 export * from "./util/IAspectExports";
-export * from "./util/ILogTarget";
 export * from "./util/IPerformanceConfiguration";
 export * from "./util/ReflectedValue";
