+++ packages/core/src/reporter/EmptyReporter.ts
@@ -1,18 +1,14 @@
-import { TestReporter } from "../test/TestReporter";
+import { IReporter } from "./IReporter";
+import { TestContext } from "../test/TestContext";
+import { TestNode } from "../test/TestNode";
 
 /**
  * This class can be used as a stub reporter to interface with the `TestContext` in the browser.
  * It will not report any information about the tests.
  */
-export default class EmptyReporter extends TestReporter {
-  constructor(_options?: any) {
-    super();
-  }
-  onFinish(): void {}
-  onGroupFinish(): void {}
-  onGroupStart(): void {}
-  onStart(): void {}
-  onTestFinish(): void {}
-  onTestStart(): void {}
-  onTodo(): void {}
+export class EmptyReporter implements IReporter {
+  constructor(_options?: any) {}
+  onEnter(_context: TestContext, _node: TestNode): void {}
+  onExit(_context: TestContext, _node: TestNode): void {}
+  onFinish(_context: TestContext): void {}
 }
