+++ packages/core/src/reporter/IReporter.ts
@@ -1,0 +1,8 @@
+import { TestNode } from "../test/TestNode";
+import { TestContext } from "../test/TestContext";
+
+export interface IReporter {
+  onEnter(ctx: TestContext, node: TestNode): void;
+  onExit(ctx: TestContext, node: TestNode): void;
+  onFinish(ctx: TestContext): void;
+}
